Add files via upload
I have created a basic web-page using  html5 and css3 in which is a sign-up page. The register.html web-page allows user to give their details like E-mail and password. I have used jsp code to receive the data from the web-page and store it in the data base. For this, I have used My SQL workbench 8.0 , tomcat server 7.0 and jdk 8.0.